+++ packages/components/Avatar/src/constants.ts
@@ -1,0 +1,4 @@
+import type { InjectionKey } from 'vue'
+import type { AvatarGroupProps } from './types'
+
+export const avatarGroupKey: InjectionKey<AvatarGroupProps> = Symbol('avatarGroupKey')