+++ packages/components/Icon/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @lite-space/icon
 
+## 0.1.0
+
+### Minor Changes
+
+- feat icon user
+
 ## 0.0.3
 
 ### Patch Changes