+++ packages/theme/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @lite-space/theme
 
+## 0.2.0
+
+### Minor Changes
+
+- feat component avatar
+
 ## 0.1.3
 
 ### Patch Changes